SLed fixed LED indexes

+++ spi_test/StateEnvAnalize.cpp
@@ -31,7 +31,7 @@ void StateEnvAnalize::OnEntry(){
     logger::log(logger::LLOG::DEBUG, TAG, std::string(__func__) + " StateEnvAnalize started");
 
     auto ctrl = get_item<pirobot::item::sledctrl::SLedCtrl>("SLedCtrl");
-    ctrl->set_color(0, 0x00334455);
+    ctrl->set_color(0, (uint32_t)0x00334455);
 
     ctrl->On();
     ctrl->refresh();
@@ -53,7 +53,7 @@ bool StateEnvAnalize::OnTimer(const int id){
 
         case TIMER_USER_1:
             {
-                ctrl->set_color(0, 0x005522AA);
+                ctrl->set_color(0, (uint32_t)0x005522AA);
                 ctrl->refresh();
 
                 logger::log(logger::LLOG::DEBUG, TAG, std::string(__func__) + " Run stop timer 5 sec");
